chore: rename bold upgrade common
