+++ scripts/boldUpgradeCommon.ts
