+++ scripts/testLocalExecuteBoldUpgrade.ts
@@ -1,6 +1,6 @@
 import { Contract, ContractReceipt } from 'ethers'
 import { ethers } from 'hardhat'
-import { Config, DeployedContracts, getConfig, getJsonFile } from './common'
+import { Config, DeployedContracts, getConfig, getJsonFile } from './boldUpgradeCommon'
 import { BOLDUpgradeAction__factory, EdgeChallengeManager__factory, RollupUserLogic__factory } from '../build/types'
 import { abi as UpgradeExecutorAbi } from './files/UpgradeExecutor.json'
 import dotenv from 'dotenv'
